feat: rename ris app module to models
