+++ deployment/dev.yaml
@@ -31,11 +31,11 @@ services:
       - SECRET_KEY=${DJANGO_SECRET_KEY}
       - SOLR_HOST=http://solr:8983/solr
       - SOLR_COLLECTION=ris
-      - RIS_DB_HOST=scraper-database
-      - RIS_DB_PORT=5432
-      - RIS_DB_NAME=${SCRAPER_DB}
-      - RIS_DB_USER=${SCRAPER_DB_USER}
-      - RIS_DB_PASSWORD=${SCRAPER_DB_PASSWORD}
+      - SCRAPED_DB_HOST=scraper-database
+      - SCRAPED_DB_PORT=5432
+      - SCRAPED_DB_NAME=${SCRAPER_DB}
+      - SCRAPED_DB_USER=${SCRAPER_DB_USER}
+      - SCRAPED_DB_PASSWORD=${SCRAPER_DB_PASSWORD}
       - DEBUG=${DEBUG}
       - UPDATE_SOLR_CRON=${UPDATE_SOLR_CRON}
       - ALLOWED_HOSTS=${ALLOWED_HOSTS}
