+++ parliscope/frontend/processing/processing.py
@@ -2,7 +2,7 @@ import re
 from datetime import datetime
 from typing import Any, cast
 
-from ris.models import AgendaItem, Consultation, Document, Meeting, Organization
+from models.models import AgendaItem, Consultation, Document, Meeting, Organization
 
 SOLR_DATE_FORMAT = "%Y-%m-%dT%H:%M:%SZ"
